move package functionality into package struct
