+++ src/rendering/scroll_selector.rs
@@ -2,8 +2,8 @@ use color_eyre::eyre::ContextCompat;
 use color_eyre::Result;
 use console::style;
 
-use crate::parsing::package::Package;
-use crate::project::dependencies::Dependency;
+use crate::project::dependency::Dependency;
+use crate::project::package::Package;
 use crate::rendering::search::highlight_search;
 
 pub struct ScrollSelector {
